Move timer code into own module
